requirements.txt
langchain==0.2.5
langchain_community==0.2.5
langchain_core==0.2.7
langchain_postgres==0.0.7
python-dotenv==1.0.1
